+++ CHANGES.rst
@@ -4,7 +4,17 @@ Changelog
 1.0.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+king changes:
+
+- *add item here*
+
+New features:
+
+- *add item here*
+
+Bug fixes:
+
+- *add item here*
 
 
 1.0.2 (2017-02-22)